move utils file
